pyproject.toml
[project]
name = "delivery-service"
version = "0.1.0"
description = ""
authors = [{ name = "DarkShine666", email = "forizn00@gmail.com" }]
readme = "README.md"
requires-python = ">3.12, <4.0"
dependencies = [
    "django (>=5.2,<6.0)",
    "djangorestframework (>=3.16.0,<4.0.0)",
    "drf-spectacular (>=0.28.0,<0.29.0)",
    "python-dotenv (>=1.1.0,<2.0.0)",
    "mysqlclient (>=2.2.7,<3.0.0)",
    "celery (>=5.5.1,<6.0.0)",
    "redis (>=5.2.1,<6.0.0)",
    "dynaconf (>=3.2.10,<4.0.0)",
    "requests (>=2.32.3,<3.0.0)",

]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
pytest = "^7.0"
pytest-django = "^4.0"
